:recycle: TodoItemViewクラスを抽出
