+++ jsprimer.net/todoapp/src/App.js
@@ -1,5 +1,6 @@
 import { TodoListModel } from "./model/TodoListModel.js";
 import { TodoItemModel } from "./model/TodoItemModel.js";
+import { TodoItemView } from "./view/TodoItemView.js";
 import { element, render } from "./view/html-util.js";
 
 export class App {
@@ -65,7 +66,8 @@ export class App {
     const todoListElement = element`<ul />`;
     // 各TodoItemモデルに対応したHTML要素を作成し、リスト要素へ追加する
     todoItems.forEach(todoItem => {
-      const todoItemElement = this.createItemElement(todoItem, {
+      const todoItemView = new TodoItemView();
+      const todoItemElement = todoItemView.createElement(todoItem, {
         onDeleteTodo,
         onUpdateTodo
       });
@@ -74,40 +76,5 @@ export class App {
     return todoListElement;
   }
 
-  /**
-   * `item`に対応するTodoアイテムのHTML要素を作成して返す
-   * @param {TodoItemModel} item
-   * @param {function({id:string, completed: boolean})} onUpdateTodo チェックボックスの更新イベントリスナー
-   * @param {function({id:string})} onDeleteTodo 削除ボタンのクリックイベントリスナー
-   * @returns {Element}
-   */
-  createItemElement(item, { onUpdateTodo, onDeleteTodo }) {
-    const todoItemElement = item.completed
-          ? element`<li><input type="checkbox" class="checkbox" checked>
-                <s>${item.title}</s>
-                <button class="delete">x</button>
-            </li>`
-          : element`<li><input type="checkbox" class="checkbox">
-                ${item.title}
-                <button class="delete">x</button>
-            </li>`;
-    const inputCheckboxElement = todoItemElement.querySelector(".checkbox");
-    inputCheckboxElement.addEventListener("change", () => {
-      // コールバック関数に変更
-      onUpdateTodo({
-        id: item.id,
-        completed: item.completed
-      });
-    });
-    const deleteButtonElement = todoItemElement.querySelector(".delete");
-    deleteButtonElement.addEventListener("click", () => {
-      // コールバック関数に変更
-      onDeleteTodo({
-        id: item.id
-      });
-    });
-    // 作成したTodoアイテムのHTML要素を返す
-    return todoItemElement;
-  }
 
 }
